// src/components/models/Typography.ts
export type TypographyVariant =
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "h6"
  | "body-large"
  | "body-large-public-sans"
  | "body-medium"
  | "body-small"
  | "body-mini"
  | "overline"
  | "button"
  | "caption"
  | "alert-message"
  | "logo-title"
  | "footer-title"
  | "bullet-list"
  | "body-mini-bold";